pyproject.toml
[tool.poetry]
name = "youtube-scraper"
version = "0.1.0"
description = ""
authors = ["Lokixin <dimasavilamartinez@gmail.com>"]
readme = "README.md"
packages = [{include="youtube_scraper", from="src"}]

[tool.poetry.dependencies]
python = "^3.11"
youtube-comment-downloader = "^0.1.70"
asyncer = "^0.0.5"
pandas = "^2.2.1"
aiofile = "^3.8.8"
psycopg = {extras = ["binary"], version = "^3.1.18"}
kafka-python = "^2.0.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.1"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
